Move length validator
